+++ libtiledbsoma/src/soma/soma_dense_ndarray.h
@@ -15,6 +15,7 @@
 
 #include <filesystem>
 
+#include "../tiledb_adapter/platform_config.h"
 #include "soma_array.h"
 
 namespace tiledbsoma {