+++ apis/r/NEWS.md
@@ -7,6 +7,7 @@
 ## Changed
 
 - Update [TileDB core to 2.28.1](https://github.com/TileDB-Inc/TileDB/releases/tag/2.28.1). ([#4077](https://github.com/single-cell-data/TileDB-SOMA/pull/4177))
+- DataFrame columns of dictionary type, with a `large_string` or `large_binary` value type, were incorrectly reported as an Arrow `string`. They are now correctly reported as dictionary-typed fields with a value type of `large_string` and `large_binary`, respectively. NB: all string/binary types are automatically up-cast to their large variant in tiledbsoma. ([#4209](https://github.com/single-cell-data/TileDB-SOMA/pull/4209))
 
 ## Deprecated